requirements.txt
aiogram==3.3.0
Django==5.0.1
djangorestframework==3.14.0
uvicorn==0.25.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2==2.9.9
drf-spectacular==0.27.0
fuzzywuzzy==0.18.0
django-filter==23.5
redis==5.0.1